スケジュールにアサイン済みのタスクが削除できないように修正

+++ app/controllers/api/tasks_controller.rb
@@ -36,6 +36,10 @@ class Api::TasksController < ApplicationController
   end
 
   def destroy
+    if Schedule.exists?(task_id: @task.id)
+      @task.errors.add(:base, "そのタスクはスケジュールにアサインされているため削除できません。")
+      return render json: @task.errors, status: :unprocessable_entity
+    end
     if @task.destroy 
       render status: 200, json: {status: 200}
     end
